Documents/Projects/smart-farming-advisor/requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
joblib==1.3.2

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# LLM & RAG (REQUIRED FOR PROJECT - Not for data download)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
google-generativeai==0.3.2
langchain-google-genai==0.0.6


# Google Gemini API (REQUIRED FOR RAG QUESTION ANSWERING)
google-generativeai>=0.4.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
requests>=2.31.0
matplotlib==3.7.2
seaborn==0.12.2

# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Evaluation metrics
nltk>=3.8.0

# Data
kaggle==1.5.16